stations: fix name/meta behavior
See the comment in the code about the discrepancy between "name" and
"meta". Interestingly, up until a while ago the API actually tolerated
these wrong requests with only "meta" set. Now, when we send a search
request with an empty "name", it returns no results (and doesn't even
look at the "meta" field).