refactoring: merge oebb-cli code base with this
Splitting a CLI app and a library into multiple projects is not really
how it's supposed to be done in the Go world ("it seemed like a good
idea at the time").

So, get rid of the oebb-cli repository and put its code into a separate
package here. Or, more accurately, replace this by the code in oebb-cli
and place the library code into a separate package ("client").